urbanwatch-ai/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
numpy==2.2.6
Pillow==11.2.1
pydantic==2.7.1
pydantic-settings==2.2.1
aiofiles==23.2.1
python-multipart==0.0.9
scipy==1.15.3

# Optional — install separately if needed:
# opencv-python-headless==4.9.0.80
# torch==2.3.0
# torchvision==0.18.0
# earthengine-api==0.1.404
# sentinelhub==3.10.2
